commented base page and removed page items
